460 Dependency Resolver with graphql: cleanup
- Removed some code from the old DependencyResolver which is no longer used. Also added ObsoleteAttribute.

- Moved PackageCompatibilityHelper out and added a comment.
